Dodawanie zespołu

+++ src/main/java/pl/kantoch/dawid/magit/models/OrganisationRole.java
@@ -11,7 +11,7 @@ public class OrganisationRole
 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(length = 40)
     private String name;
@@ -22,7 +22,7 @@ public class OrganisationRole
     @Column(name = "organisation_id")
     private Long organisationId;
 
-    public OrganisationRole(Integer id, String name, String iconName, Long organisationId) {
+    public OrganisationRole(Long id, String name, String iconName, Long organisationId) {
         this.id = id;
         this.name = name;
         this.iconName = iconName;
@@ -32,7 +32,7 @@ public class OrganisationRole
     public OrganisationRole() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -44,7 +44,7 @@ public class OrganisationRole
         this.iconName = iconName;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
